Factor out static-page.tsx.
